Change load package function to ensure_packages
Vibe coded with claude on this one :(
Had an RENV set up but it was too complicated and hard to maintain. This function checks for any packages that might exist in a users library path and if it is missing packages, it will create a local path and a folder in functions that contains all the missing packages and dependencies. It will then append that library path to the defualt library path for R, so that R will know to look there.
Finally, tidyverse just loads WAYY too many things. I went through and looked at what you actually use from tidyverse and its just a few packages. So i updated the other functions to only load those pacakges.
i know its less simple (one package vs a couple now), but installing tidyverse was taking forever on my end.